Renamed rails helper to spec helper
