+++ app/controllers/application_controller.rb
@@ -1,3 +1,3 @@
-class ApplicationController < ActionController::Base
-  self.abstract_class = true
+class ApplicationController < Sinatra::Base
+  set root: File.expand_path("..", __dir__)
 end
