+++ config.ru
@@ -7,4 +7,4 @@ if ActiveRecord::Base.connection.migration_context.needs_migration?
 end
 
 run ApplicationController
-# use MerchantsBusinessLogicController
+ use RecordingsController
