+++ spec/features/api/v1/recording/recording_request_spec.rb
@@ -1,11 +1,11 @@
-require "./spec/rails_helper"
-require "./app/controllers/application_controller"
+require "./spec/spec_helper"
+# require "./app/controllers/application_controller"
 
-describe "Vote Requests" do
+describe "Vote Requests", type: :request do
   it "Gets ratings for a specific location" do
 
-    get '/api/v1/Recording/1/recording_rating'
-
+    get '/api/v1/recording/1/recording_rating'
+binding.pry
     expect(last_response).to be_successful
 
     ratings = JSON.parse(last_response.body)
