+++ spec/models/recording.rb
@@ -1,0 +1,6 @@
+class Recording
+  
+  def initialize(attributes)
+    @title = attributes
+  end
+end
